tender-automation/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
python-dotenv==1.1.1
openai==1.99.1
pypdf==5.9.0
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==11.3.0
